# src/config/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"SuccessCode": {
				"enum": [
					"OPERATION_SUCCESS"
				],
				"type": "string"
			},
			"GeneralResponse": {
				"properties": {
					"data": {},
					"code": {
						"$ref": "#/components/schemas/SuccessCode"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"data",
					"code",
					"message"
				],
				"type": "object"
			},
			"_36_Enums.Gender": {
				"type": "string",
				"enum": [
					"male",
					"female"
				]
			},
			"Gender": {
				"$ref": "#/components/schemas/_36_Enums.Gender"
			},
			"_36_Enums.UserRole": {
				"type": "string",
				"enum": [
					"admin",
					"user"
				]
			},
			"UserRole": {
				"$ref": "#/components/schemas/_36_Enums.UserRole"
			},
			"CreateUserDTO": {
				"properties": {
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"age": {
						"type": "number",
						"format": "double"
					},
					"gender": {
						"$ref": "#/components/schemas/Gender"
					},
					"avatar": {
						"type": "string"
					},
					"role": {
						"$ref": "#/components/schemas/UserRole"
					},
					"password": {
						"type": "string"
					},
					"confirmPassword": {
						"type": "string"
					}
				},
				"required": [
					"firstName",
					"lastName",
					"email",
					"phone",
					"age",
					"gender",
					"role",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateUserDTO": {
				"properties": {
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"age": {
						"type": "number",
						"format": "double"
					},
					"gender": {
						"$ref": "#/components/schemas/Gender"
					},
					"avatar": {
						"type": "string"
					}
				},
				"required": [
					"firstName",
					"lastName",
					"phone",
					"age",
					"gender"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AuthRequest": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "badmintonapi",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/badminton/v1/users": {
			"get": {
				"operationId": "GetUsers",
				"responses": {
					"200": {
						"description": "Thông tin người dùng.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GeneralResponse"
								}
							}
						}
					}
				},
				"description": "Lấy thông tin người dùng theo ID.",
				"tags": [
					"User Management"
				],
				"security": [
					{
						"jwt": [
							"ADMIN"
						]
					}
				],
				"parameters": [
					{
						"description": "ID của người dùng.",
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "Thông tin người dùng sau khi cập nhật.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GeneralResponse"
								}
							}
						}
					}
				},
				"description": "Cập nhật thông tin người dùng (chỉ Admin).",
				"tags": [
					"User Management"
				],
				"security": [
					{
						"jwt": [
							"ADMIN"
						]
					}
				],
				"parameters": [
					{
						"description": "ID của người dùng.",
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"description": "Dữ liệu mới cần cập nhật.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateUserDTO",
								"description": "Dữ liệu mới cần cập nhật."
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"200": {
						"description": "Kết quả xóa người dùng.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GeneralResponse"
								}
							}
						}
					}
				},
				"description": "Xóa người dùng theo ID (chỉ Admin).",
				"tags": [
					"User Management"
				],
				"security": [
					{
						"jwt": [
							"ADMIN"
						]
					}
				],
				"parameters": [
					{
						"description": "ID của người dùng cần xóa.",
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/badminton/v1/users/email/{email}": {
			"get": {
				"operationId": "GetUsersByEmail",
				"responses": {
					"200": {
						"description": "Thông tin người dùng.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GeneralResponse"
								}
							}
						}
					}
				},
				"description": "Lấy thông tin người dùng theo email.",
				"tags": [
					"User Management"
				],
				"security": [
					{
						"jwt": [
							"ADMIN"
						]
					}
				],
				"parameters": [
					{
						"description": "Email của người dùng.",
						"in": "path",
						"name": "email",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/badminton/v1/users/all": {
			"get": {
				"operationId": "GetAllUsers",
				"responses": {
					"200": {
						"description": "Danh sách người dùng với phân trang.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GeneralResponse"
								}
							}
						}
					}
				},
				"description": "Lấy danh sách tất cả người dùng với phân trang.",
				"tags": [
					"User Management"
				],
				"security": [
					{
						"jwt": [
							"ADMIN"
						]
					}
				],
				"parameters": [
					{
						"description": "Chỉ số trang.",
						"in": "query",
						"name": "pageIndex",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "Kích thước trang.",
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/badminton/v1/users/signUpInternal": {
			"post": {
				"operationId": "CreateUserInternal",
				"responses": {
					"200": {
						"description": "Thông tin người dùng vừa tạo.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GeneralResponse"
								}
							}
						}
					}
				},
				"description": "Đăng ký tài khoản mới (chỉ admin)",
				"tags": [
					"User Management"
				],
				"security": [
					{
						"jwt": [
							"ADMIN"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "Dữ liệu người dùng cần tạo.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateUserDTO",
								"description": "Dữ liệu người dùng cần tạo."
							}
						}
					}
				}
			}
		},
		"/badminton/v1/users/signUp": {
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"200": {
						"description": "Thông tin người dùng vừa tạo.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GeneralResponse"
								}
							}
						}
					}
				},
				"description": "Đăng ký tài khoản mới (auto role: user)",
				"tags": [
					"User Management"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "Dữ liệu người dùng cần tạo.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateUserDTO",
								"description": "Dữ liệu người dùng cần tạo."
							}
						}
					}
				}
			}
		},
		"/badminton/v1/users/me": {
			"put": {
				"operationId": "UpdateCurrentUser",
				"responses": {
					"200": {
						"description": "Thông tin người dùng sau khi cập nhật.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GeneralResponse"
								}
							}
						}
					}
				},
				"description": "Cập nhật thông tin của bản thân.",
				"tags": [
					"User Management"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "Dữ liệu mới cần cập nhật.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateUserDTO",
								"description": "Dữ liệu mới cần cập nhật."
							}
						}
					}
				}
			}
		},
		"/badminton/v1/auth/user": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Thông tin xác thực và token.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GeneralResponse"
								}
							}
						}
					}
				},
				"description": "Đăng nhập người dùng.",
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "Thông tin đăng nhập.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AuthRequest",
								"description": "Thông tin đăng nhập."
							}
						}
					}
				}
			}
		},
		"/badminton/v1/auth/me": {
			"get": {
				"operationId": "GetCurrentLogin",
				"responses": {
					"200": {
						"description": "Thông tin người dùng.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GeneralResponse"
								}
							}
						}
					}
				},
				"description": "Lấy thông tin người dùng hiện tại.",
				"tags": [
					"Authentication"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}